// shared/types/enums/OrderStatus.d.ts
export declare enum OrderStatus {
    PENDING = "pending",
    CONSOLIDATED = "consolidated",
    ASSIGNED = "assigned",
    ROUTED = "routed",
    DISPATCHED = "dispatched",
    IN_TRANSIT = "in_transit",
    DELIVERED = "delivered",
    FAILED = "failed",
    CANCELLED = "cancelled"
}
//# sourceMappingURL=OrderStatus.d.ts.map